fix: unificar proxy.ts para evitar conflicto de matcher
Next.js 16 descubría dos proxy.ts (raíz y src/) y usaba el matcher
del src/proxy.ts, que no excluía imágenes. Esto hacía que next-intl
interceptara todas las peticiones de imágenes añadiendo prefijo de
idioma (/es/...), devolviendo 404.

- Fusiona lógica de staging/auth + redirects + next-intl en proxy.ts raíz
- Elimina src/proxy.ts para evitar conflicto de matcher
- Mantiene matcher correcto: excluye api, _next, _vercel y archivos
  con extensión (.*\..*) incluyendo .webp, .xml, .txt

Fixes images returning 404, robots.txt returning 404, sitemap.xml
returning 404, and API routes being rewritten with locale prefix.
